Splitting bin and lib logic
