+++ src/commands/init.rs
@@ -18,6 +18,7 @@ impl Init {
 impl Command for Init {
     fn execute(&mut self) -> Result<(), RitError> {
         let git_path = self.git_path();
+        dbg!(&git_path);
 
         for dir in &["objects", "refs"] {
             fs::create_dir_all(git_path.join(dir))?;
