+++ src/commands/mod.rs
@@ -1,5 +1,8 @@
-use crate::{errors::RitError, Session};
-use std::path::PathBuf;
+use crate::errors::RitError;
+use std::{
+    io::{self, Read},
+    path::PathBuf,
+};
 
 mod init;
 use init::Init;
@@ -9,6 +12,29 @@ use commit::Commit;
 
 mod add;
 use add::Add;
+
+#[derive(Clone, Debug)]
+pub struct Session {
+    pub name: String,
+    pub email: String,
+}
+
+impl Session {
+    pub fn new(name: Option<String>, email: Option<String>) -> Result<Self, RitError> {
+        let name = name.unwrap();
+        let email = email.unwrap();
+
+        Ok(Self { name, email })
+    }
+
+    pub fn read_stdin(&self) -> Result<String, RitError> {
+        let mut text = String::new();
+
+        io::stdin().read_to_string(&mut text)?;
+
+        Ok(text)
+    }
+}
 
 trait Command {
     fn execute(&mut self) -> Result<(), RitError>;
