+++ src/lib.rs
@@ -1,0 +1,21 @@
+mod objects;
+
+mod workspace;
+
+mod database;
+
+mod refs;
+
+pub mod lockfile;
+
+pub mod index;
+
+mod id;
+
+mod repository;
+
+mod commands;
+
+pub use commands::*;
+
+pub mod errors;
