+++ src/main.rs
@@ -1,67 +1,20 @@
-use commands::CommandOpts;
-use io::Read;
-use lockfile::LockError;
-use std::{env, io};
+use std::env;
 
-mod objects;
-
-mod workspace;
-
-mod database;
-
-mod errors;
-use errors::RitError;
-
-mod refs;
-
-mod lockfile;
-
-mod index;
-use index::IndexError;
-
-mod id;
-
-mod repository;
-
-mod commands;
-
-pub struct Session {
-    pub name: String,
-    pub email: String,
-}
-
-impl Session {
-    fn new() -> Result<Self, RitError> {
-        let name = env::var("GIT_AUTHOR_NAME")?;
-        let email = env::var("GIT_AUTHOR_EMAIL")?;
-
-        Ok(Self { name, email })
-    }
-
-    pub fn read_stdin(&self) -> Result<String, RitError> {
-        let mut text = String::new();
-
-        io::stdin().read_to_string(&mut text)?;
-
-        Ok(text)
-    }
-}
-
-fn exit(result: Result<(), RitError>) {
+fn exit(result: Result<(), rit::errors::RitError>) {
     std::process::exit(match result {
         Ok(_) => 0,
         Err(err) => match err {
-            RitError::MissingFile(_) => {
+            rit::errors::RitError::MissingFile(_) => {
                 eprintln!("fatal: {}", err);
                 128
             }
-            RitError::PermissionDenied(_) => {
+            rit::errors::RitError::PermissionDenied(_) => {
                 eprintln!("error: {}", err);
                 eprintln!("fatal: adding files failed");
                 128
             }
-            RitError::Lock(err) => match err {
-                LockError::Denied(_) => {
+            rit::errors::RitError::Lock(err) => match err {
+                rit::lockfile::LockError::Denied(_) => {
                     eprintln!("fatal: {}", err);
                     128
                 }
@@ -70,8 +23,8 @@ fn exit(result: Result<(), RitError>) {
                     1
                 }
             },
-            RitError::Index(err) => match err {
-                IndexError::Lock(lock_err @ LockError::Denied(_)) => {
+            rit::errors::RitError::Index(err) => match err {
+                rit::index::IndexError::Lock(lock_err @ rit::lockfile::LockError::Denied(_)) => {
                     eprintln!(
                         "fatal: {}
                         
@@ -102,9 +55,12 @@ fn main() {
 
     let args = args.collect();
     let dir = env::current_dir().unwrap();
-    let session = Session::new().unwrap();
 
-    let result = commands::execute(CommandOpts { dir, session, args });
+    let name = env::var("GIT_AUTHOR_NAME").ok();
+    let email = env::var("GIT_AUTHOR_EMAIL").ok();
+    let session = rit::Session::new(name, email).unwrap();
+
+    let result = rit::execute(rit::CommandOpts { dir, session, args });
 
     exit(result);
 }
